Refactor: add the assign_utils for common assign methods
